Use merge strategy for mergeability testing on "Rebase if Necessary" strategy
When a user pushes a chain of commits, they typically have inner
dependencies, and it is likely that a commit in the middle of the
chain cannot be merged by cherry-picking it alone, but if the chain
is submitted in order it will be merged just fine.

Applying cherry-pick for mergeability test fails for all the commits
that cannot be picked without their dependencies, but this is obviously
wrong when using "Rebase if Necessary", since the changes are never
applied without their dependencies.

With gerrit 2.12 this is even worse, since the Submit button is disabled
for all the changes that are wrongfully marked as conflicted.

Bug: Issue 3742
Change-Id: I6dd4d53e663bc9960c4c64b05cd011b87edcdabb

Former-commit-id: f45ace6bc360fa4b736e0665439f7d827675b4f6