MergeValidationListener: Expose the caller in onPreMerge() method
After recent merge queue refactoring, there is no way to retrieve
submitter in MergeValidationListener#onPreMerge() method any more.
Expose the caller so that validator implementations can implement
site specific checks. Note, that even when there would be a way to
perform database and cache lookups and retrieve the submitter, this
was always wrong API design: instead of forcing the plugin authors
to retrieve the caller and buy the performance degradation caused
by peforming database and cache lookups, the caller can be easily
passed in.

This change breaks some plugins, that implement the listener, e.g.
cookbook-plugin. However, given that there is no alternative way
to retrieve the caller, this API change is justified.

Reported-by: Orgad Shaneh <orgads@gmail.com>
Change-Id: I7cb5e354a67de4f21bb4eee6d53a04a18624cc2c

Former-commit-id: 84bc19bc3d537601a7df5be670d977b113821344