Correctly detect symlinked log directory on startup
On startup the FileUtil.mkdirsOrDie() method uses the
Files.createDirectories() method to detect if the
'$site_path/logs' directory exists and create it if not. This method
fails if the '$site_path/logs' is a symlink to another directory and
prevents the server from starting. Add a call to Files.isDirectory()
first to correctly detect this scenario and allow the server to start.

Bug: Issue 3733
Change-Id: I0f7e2c94086641178c1772a404baa87f3bd7deb2

Former-commit-id: ccb946c97ae43ea83e27afcf8c268fc590178bdc