Disable submit button instead of hiding it
Meet the same logic as with old UI and disable the submit button,
when current change depends on other changes which are not ready.
Style disabled button with background-color dark grey.

Test Plan:

* Configure submit strategy "Merged if Necessary"
* Push two changes, c_1, c_2
* Vote -2 on c_1
* Vote +2 on c_2

Observe that with this change "Submit including parents" button
is shown but disabled. Hovering over the button activates this
tooltip:

  "This change depends on other changes which are not ready"

Change-Id: Icaed0db1b03a2f0dcfc0ee8f4adef62a19a2408c

Former-commit-id: a46867387402466327f5684dbcf5116b06f94c20