ChangeSet: Expose changes keyed by ID
We already build a set of these during construction time to ensure
there are no duplicates; it's not really any extra work to just
convert that set to a map and store it.

Use a LinkedHashMap for stable order, which is not especially
important for actual merge operation but does make SubmitTypeRuleIT
deterministic.

Change-Id: I2430084bef6d7b4701f3cfb6745cddb2a6d67621

Former-commit-id: b42441099cf8512a3380fec24aaba5008b96d87b