ChangeEditIT: Fix incorrect usage of ExpectedException
Having an expected exception at the beginning of a test
short circuits the assertions that follow the expect() call.

"JUnit uses AssertionErrors for indicating that a test is failing.
You have to call assert methods before you set expectations of the
ExpectedException rule, if they should be handled by the framework."[1]

Remove the assertUnchangedMessage helper method and the two calls
to it from the updateMessage test.

The first case that was being tested is already covered by the
updateMessageNoChange test.  Add a new test to cover the other
case, where new lines are added onto the edited message.

[1]: http://junit.org/apidocs/org/junit/rules/ExpectedException.html

Reported-By: Richard Ipsum <richard.ipsum@codethink.co.uk>
Change-Id: I565375da7d9511a17bbe1e146c5ab402be1a8045

Former-commit-id: 5a74e285ac36645d570396d06c28967d955a5c51