MergeOp: Simplify stamping labels at submit time
Instead of creating one ChangeUpdate per user, we can now stamp labels
for all users in a single ChangeUpdate. We still need two updates in
the case where a submit strategy produced a new patch set, but at
least the number is now bounded, so the code is simpler.

While we're in there, restructure it slightly so we don't have to run
the LabelNormalizer multiple times, and instead reuse the results from
the old patch set directly when updating the new patch set. This is
not be a behavior change, since labels are just normalized based on
the project config and permissions, which haven't changed. It just
saves a little work and makes the intent clearer.

Change-Id: I897bed1a568166ece34d9f08e27e6b219d89073a

Former-commit-id: 01cfa56a733aa5e2d6aeace056a685b8e09d8b9c