Notedb: Add a git repo based sequence implementation
This is intended to replace the next*Id() methods in ReviewDb. We
increment a counter stored in a blob on refs/sequences/*, and allow
multiple processes to reserve chunks of the number space so they can
operate concurrently. The resulting IDs are not quite monotonic, but
this is a common feature of distributed unique ID assignment; the
careful observer will note the same behavior on gerrit-review, for
example.

Change-Id: Ia5f59233df97b7eff71b1ec6af1d816975f28ad9

Former-commit-id: 9b82bfdd57f90ffe744c459e220ba12def719634