RebaseUtil: Make all methods non-static
Do depend on the injected ReviewDb provider; callers are all just
passing in their own from their own provider.

Do not depend on the injected GitRepositoryManager. In all but one
existing caller of any method in this class, there is already a repo
available. The consistency in the API is worth shuffling a bit of code
into RevisionApiImpl.

Change-Id: I3df8e7f35ce577a2ba52d60feb2cae8edd33e8cf

Former-commit-id: f85bd49cfcceb1260405ab180cbe88250a269fee