Correctly call ssh-keygen on Win32 with empty passphrase
When Gerrit is initialized with BouncyCastle libs, it creates SSH2 host key
using ssh-keygen (in other cases it is generated in java code directly).
One of the argument to ssh-keygen is empty passphrase. On Windows empty
arguments are ignored as documented in JDK-6518827. This leads to error in
ssh-keygen, surprisingly printed as "Too many arguments". The only
workaround is to manually quote the empty argument.

Change-Id: I1a03b2ddbfed700bfd885c1e4f6075f935e4bd9c

Former-commit-id: 5e8908c97286d1ae1658a61a78431f68837a0f99