RebaseIfNecessary: Don't copy approvals with ApprovalCopier
This wastes effort and introduces the constraint violation that we had
to work around and comment. Plus, there is an unfortunate restriction
in the gerrit-review gwtorm code currently where we can't modify the
same entity twice in one transaction.

Selectively bypass approval copying entirely in RebaseChangeOp, and do
it manually in RebaseIfNecessary.

Change-Id: Idd16eaef43916d0004c9622a395bd6be16fbfdd2

Former-commit-id: 3ae9aab1f2f87a81736daf5a060278c8e0eeff6c