Pass only ref name into ChangeInserter, not ref control
ChangeInserter only needs the ref control when validation is done and
it can just instantiate it for this purpose.

The project and the user are provided by the Context and a provided
ref control may have been created for another project or user. By not
passing in the ref control we avoid this possible inconsistency.

Also remove the getProject() and getUser() methods from
ChangeInserter. They were only used by tests and were needed only
because the API required to pass in the project and user twice, once
through ref control to the ChangeInserter and once to the BatchUpdate.

Change-Id: I2c4faf27e3c1183a474fecfa06dbbbc14a3c143f
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 0242ea54d3452a4c1bf37ce7cb6a47301f696230