PatchSet: Correct misspelling and type of pushCertificate field
I519240547 extended the patch_sets table with a pushCertificate column
to store push the certificate that was used when pushing a signed
commit.

However, the column type VARCHAR(255) is too small for PGP certificate
descriptions. The column name was also misspelled as "pushCertficate"
(missing a 'i').

Extend the column from VARCHAR(255) to CLOB, as used in the account
table ssh key column. At the same time, correct the misspelling.

Reported-By: Henrik Laban Sköllermark <laban@kryo.se>
Change-Id: I3dff91041c00f619f06c0e64c5cb09e3987eac9a

Former-commit-id: 547886f39671ff54150f7198d22559a0d0894896