Don't throw IOException from ChangeInserter.createChange
This IOException is only thrown when the UTF-8 encoding is not
available which is an illegal state. Hence we should rather wrap the
IOException into an IllegalStateException. This makes the method
signature cleaner and callers don't need to handle the IOException.

In JGit the ChangeIdUtil.computeChangeId method was updated so that it
no longer throws the IOException [1]. Once we upgrade to a JGit
version that contains this change we can remove the exception handling
completely.

[1] https://git.eclipse.org/r/64670

Change-Id: I2839b1f640159a30d791c5063351162f004a65b0
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: d5e930c05fef76ea3b1e305931ef16b2cd1f3ae4