ReceiveCommits: Reopen repo in background threads
JGit Repository instances are not inherently threadsafe. We already
detect the background thread case and reopen the ReviewDb; we should
also reopen the repo.

Encapsulate repo, RevWalk, and db in a new AutoCloseable,
RequestState. This can be constructed either with new objects or
reusing the ones from the current thread, allowing us to save a bit of
code at call sites.

Change-Id: I9333f68ac222c149e48680d61167d2c6261fa4f2

Former-commit-id: f277cdef1beacb03aa4bb82ba2206d242145a22f