Always advance MergeTip during updateRepo
Multiple batch update operations expect to see the results of
previous updates. In the case of inspecting the current MergeTip,
subsequent updateRepo calls expect to see the tip advancing. However,
this was not always done in updateRepo; in the rebase and cherry-pick
cases we were mistakenly delaying this until updateChange.

This resulted in a serious bug during RebaseIfNecessary, where commits
after the first one were all rebased onto the original branch tip
rather than following previous commits in the series. Worse, this case
was untested, so add a test for that.

Change-Id: Ib4f0a5667a8d2b7d37ca75b81368b24ef79374fb

Former-commit-id: 694db8ebc9c23b3172239bf70328c82b3ecd1c4a