Parallelize creating of reflog in refs/meta/config by migration step 106
The migration step creates a reflog file for the refs/meta/config
branches of all projects. This was done in a loop which caused very
slow execution when upgrading a system with a large number of projects.

Now this task is parallelized. Four threads are used per CPU as we can
expect the task to be limited by (blocking) IO.

Change-Id: I7a9c94b2f98921893226fe4c3b18a67e06f856f8

Former-commit-id: 948d12b6177001124c4b7553cb680d872bd30bb4