Send email using email queue instead of the default queue
Some emails sent asynchronously were already being sent using that queue
but some were not. This was confusing for a gerrit administrator because
if there is a build up of "send-email" tasks in the queue, he would
think that increasing sendemail.threadPoolSize would help but it did not
because some of the email were sent using the default queue which is
configurable using execution.defaultThreadPoolSize.

Change-Id: I08644be07db7455e781e36bf2b67c6d16bdd8b16

Former-commit-id: 2258fd0d40a7220b3028e2e302a8616a2ee12ab9