Fix merge validator exception type is ignored
The switch case in updateChangeStatus was supposed to handle the failed
cases but it was never reached by changes who failed to merge, since
only the submitted ones are returned by the validation function.

Now, updateChangeStatus is called on every change that doesn't make it
out of the validation. Also, the parameter name 'submitted' did not make
sense since the method handles changes who cannot be submitted, so it is
renamed to 'changes'.

Also, destBranch is optionally null in updateChangeStatus since it
doesn't apply to failed merges.

Tests are modified since the error output is more verbose than before.

Bug: Issue 3741
Change-Id: Id1de81b3136b94e210c957a58b5c801a1faa3244

Former-commit-id: dd41c53bd72b363f3f35f9a8e0dc5238c35d1b8b