XSRF token cookie should honor auth.cookieSecure setting
Change Ie24051b48 introduced sending XSRF token as a cookie. When
writing this cookie, however, the auth.cookieSecure setting in the
Gerrit configuration was not set.

In a specific setup, when the Gerrit server is behind a proxy server and
SSL termination is done at the frontend level, this broke the ability to
login directly in the backend server using the server's URL.

Fix this by using the auth.cookieSecure setting from the Gerrit
configuration when creating the cookie as it is done when creating the
GerritAccount cookie.

Change-Id: Ib17c284cf358948e9a63bf74a6d8a2519a4a4039

Former-commit-id: eca8bde3ff7481828706138fc664e229dd645156