Notedb: Move individual note parsing into a new class
We would like to store additional information besides inline comments
in notes. As a first step towards doing so, encapsulate parsing state
into a new class, RevisionNote. This class reads the note data once
and can parse it into multiple fields (but for now still one).

When updating the note, we copy the immutable RevisionNote into a
mutable RevisionNoteBuilder, which knows how to write itself out to a
note. This shifts a little bit of work from CommentsInNotesUtil to
ChangeUpdate.

Change-Id: Ia55888b043a5624b81a3fdde733363cd1f1a5aee

Former-commit-id: 3afa3a5da8dfef40f450b2ec28bcebc2e956b151