Notedb: Don't support non-UTF-8 charsets in note contents
We were doing this in a broken way, using RawParseUtils#parseEncoding,
which looks for a paragraph starting with "encoding ". This is
intended for commits and doesn't make sense in a note body.

For now assume everything is UTF-8. There is nothing in the git-notes
format that specifies encoding of note blobs, just as there is nothing
in the git format that specifies encoding of any other blobs in any
other trees. If we want to support multiple encodings later, we can
revisit this, either actually using this encoding header or using a
footer in the meta graph or something else.

Change-Id: Iaf472a27009356cfa3db9ed90e65fbc3a8b34bdb

Former-commit-id: 625a390b185356389d9fae04197ada105c58e229