NoteDb: Check change message author matches update author
At the moment the author of a change message is ignored when the
change message is written to notedb and the change message is recorded
on behalf of the user for which the ChangeUpdate was instantiated.
These users should be the same, but add a check to verify this so that
the change message data in the database and in notedb cannot become
inconsistent.

Change-Id: I6132a2dda18b5cfb27563211ebf9fc70dbef403d
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 917fa77ca1d23825b46cb5686f6b1a903c9550d3