Configure editable account fields for OAuth realm
Full name and email address might not be provided by all OAuth login
providers. In that case, it should be possible for a user to enter
these account fields in the Gerrit UI manually.

OAuthRealm now checks the Gerrit configuration for the parameters
oauth.allowEditFullName and oauth.allowRegisterNewEmail. If set to
true, a user may edit the full name or add additional email
addresses in the Contact Information settings, respectively.
Otherwise Gerrit relies on the information provided by the OAuth
login provider.

Change-Id: I250e86ada75e1d0a8e1018903510e2d1b55b8fa9
Signed-off-by: Michael Ochmann <michael.ochmann@sap.com>

Former-commit-id: 6981326d6981b2471acae1b08886eb41487c3e82