ChangeJson: Fix approvals for draft changes
In If07a6e67 we started caching the submit records produced by MergeOp
back in the ChangeData. An unintended side effect of this was to reuse
those cached submit records in ChangeJson, since MergeOp is called
when setting the submittable bit. The SubmitRuleEvaluator in MergeOp
runs with allowDrafts = false, meaning no submit records would appear
later in ChangeJson when we run initLabels on draft changes. Another
side effect is making the reviewer list empty on the change screen,
which still uses the labels map rather than the reviewers map.

Fix this in a slightly hacky way by using a cache specific to
ChangeJson which guarantees that the right value of allowDrafts is
set.

Change-Id: Ifb95f8796388facb05cddb2870f1d6d649e25088

Former-commit-id: 5abe28c4f9bbd98fc41fbe2e5adc6631f12ac629