Update Buck to newest version
Now, that cross-cell support is implemented: [1] we can share rules
between projects. Cell is a container with its own buck-out directory.
The Buck design allows to replace a cell with alternative location.
This is a preparation change to support cross-cell dependencies.

The prerequisite for the hijacked cell to work properly, is that the
classpaths are always represented as absolute paths. That was not the
case in Buck implementation and thus classpath resolution across cell
boundaries was broken. This got fixed in: [2], [3].

[1] https://github.com/facebook/buck/issues/116
[2] https://github.com/facebook/buck/issues/545
[3] https://github.com/facebook/buck/pull/558

Change-Id: I3b17129c0f7ca5801f03091fff5651fb63d1482f

Former-commit-id: feeb212308d935642c681f1f1e367535f5c5d6a0