Kill CommentsTest
Maintaining this test became a burden for developing new features in
Gerrit. Its complicated mock setup must be adapted whenever a new
dependency is introduced for any of the involved classes. Instead of
wasting more time with keeping this test working, extend the
CommentsIT integration test to cover the functionality that was tested
by CommentsTest:

- CommentsTest#testListComments():
    Replaced by CommentsIT#listComments()

- CommentsTest#testGetCommentExisting():
    Replaced by an additional assert in CommentsIT#postComment()

- CommentsTest#testGetCommentNotExisting():
    Replaced by CommentsIT#getNonExistingComment()

- CommentsTest#testListDrafts():
    Replaced by CommentsIT#listDrafts()

- CommentsTest#testPatchLineCommentsUtilByCommentStatus() &&
  CommentsTest#testPatchLineCommentsUtilDraftByChangeAuthor():
    Covered by the sum of the tests in CommentsIT

Change-Id: I5b898645080974db4eb3d511d834383675db1e06
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 4ed523d0e6602b8c1e21b3028cdbae425b459f1f