Merge CheckIT into ConsistencyCheckerIT
These were originally split when ConsistencyCheckerTest was not an
acceptance test. Now that it is, there's no reason to have two.

The only case that was not already tested in ConsistencyCheckIT was
that the result ChangeInfo was properly updated after fixing. Write a
simpler test for this than what we had before.

Change-Id: Ib88442cf062d032104d7a0e35d9348ba5c103246

Former-commit-id: 1c7aa94de4354c6cc4ca4a1206bc49c9994385fa