Notedb: Fully implement deleting draft changes
Rather than try to work within the strictures of VersionedMetaData,
dedicate a new class ChangeDelete to do this with a simple series of
ref updates.

Although in theory we could do what we do for patch sets and add a
DELETED state to changes, that would be too much of a behavior change
from the existing behavior. We had to do it in the case of patch sets
to avoid rewriting history, but we sidestep that issue here because
deleting refs is not really rewriting history.

Change-Id: Ib200aeea160921dbc9181e0341f3542f05651577

Former-commit-id: b3d7f7328e03b924d4236ff27b0425b90767758e