Fix hashtag set operations
The current hashtag set can never be null, but it may be empty. When
empty we still need to prune hashtags from the toRemove set, and the
logic in this block otherwise works when the set is empty, so just get
rid of the conditional.

HashtagsIT was intending to check for this, but it didn't work since
we were comparing message timestamps while also using the system
clock. Change to use a testing clock, and also check message contents
as well so the failure is more obvious.

Change-Id: I6bd852fc357e99bc42c53db75fa690930e1e3ed0

Former-commit-id: 022ab1291c5d947d1cef77d9e935e2f9082ff797