Allow uploader to control who should be notified on upload
A new 'notify' option is added that can be specified on push:

  git push ssh://bot@gerrit:29418/foo HEAD:refs/for/master%notify=NONE

It supports the same values as the notify option in ReviewInput:

  NONE, OWNER, OWNER_REVIEWERS, ALL

This is useful to reduce email noise when bot users are automatically
uploading changes and patch sets. E.g. a bot that is able to forge
committer and author may want to upload a commit without cc-ing the
author and the committer of that commit.

Change-Id: I473703a439018c3097f5c3de8f47d3f8371551ab
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 9e078d805b56443181cd703b9e9efa058721d07f