MergeOp: Always retain canMergeFlag on reset
Methods in places like MergeUtil often want to reuse the RevWalk to
walk a little segment of history. To retain the canMergeFlag on the
walk, they currently have to have the flag argument passed in and then
call resetRetain. Unfortunately this complicates method signatures and
is easily forgotten.

Fortunately RevWalk has a retainOnReset method that can be used to
always retain a given flag even if downstream consumers forget to call
resetRetain. Use that to simplify some code.

There are still a few helper methods that actually use the
canMergeFlag in their computations, so they still get it as an
argument, and I left the resetRetain calls in there to make it clear
that retaining the flag is specifically intended in these cases.

Change-Id: Ib36ba8bafd701442bd078d671650e0bebe15529d

Former-commit-id: 5ac89f365a2a26dc8d5df37c71ca1268d6a66fb8