Old DownloadCommand class must be used in intermediate upgrade phase
To support zero downtime upgrade, move of user preferences data from
database to git backend is implemented in two phases. In second phase
the class AccountGeneralPreferences is dropped and replaced with
GeneralPreferencesInfo class. In DownloadConfig this replacement was
erroneously done already in the first phase: I218d9246. This broke
download-commands plugin. Restore the old class and postpose it
replacement in the phase 2 of migration.

Bug: Issue 3858
Change-Id: I70dbf4ce39239871fc2f67161f22eb402a83b2f5

Former-commit-id: f8fb7f11eb7c8c0dc967641b4a0aabdc22b4e4c6