Always load change notes when they are created
This is a preparation for when reviewdb is gone and changes are fully
initialised from notedb. Then we will instantiate the Change object
only in ChangeNotes#onLoad, and no longer in
ChangeNotes.Factory#create. This is because the Change can only be
instantiated after parsing the notes branch. The problem is that lots
of callers access the Change object by ChangeNotes#getChange() before
invoking ChangeNotes#load() in which case the change would not be
available yet. Instead of adapting all callers to invoke
ChangeNotes#load() before asking for the change, make sure that the
change notes are always loaded on creation.

Change notes are still not loaded eagerly when they are created from
an indexed change. This is because we don't want to access notedb
during a query (similar to how we don't want to access the database
during a query).

Change-Id: Id654f192910e131d695e5ba0d60daf18bf736ded
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: d4038d6e1b2a42b4f75ffc529be274bbd9f61fcc