RegisterNewEmailSender: Don't unconditionally send verification mail
If the server configuration includes sendemail.allowrcpt, and the
user attempts to register an email address that is not included in
the allowed list, OutgoingEmail will skip adding that address to the
list of recipients.  However, RegisterNewEmailSender overrides the
'shouldSendMessage' method to always return true, which means Gerrit
still attempts to send the verification mail which has no recipients.

This results in an error from the SMTP server:

   rejected DATA command: 503 5.5.2 Need rcpt command

which in turn results in an "Internal server error" message to the
user.

Remove the overridden shouldSendMessage method, so that the default
implementation (which checks for recipients and content) is used.

Change-Id: I4f756f98627447a3f23ecdc9fa82309ba0a0271a

Former-commit-id: e8f99813a7b0f4e2fa9ac0c48702d26047c4d95b