OutgoingEmail: Check for valid email address when adding recipients
After finding error logs filled with the message:

  Not emailing NULL (prohibited by allowrcpt)

it turned out that a user's preferred email address had been set to
the literal string "NULL". Presumably this was done while running a
version of Gerrit older than 2.9, which included change I1f8d95dd9
("Validate email address when adding email or creating account").

In combination with the allowrcpt setting being enabled, this was
preventing the user from receiving any mails from Gerrit, and causing
the log to be filled with this message.

When adding recipients to an outgoing mail, check for the email address
being invalid, and log a different message in this case.

Change-Id: Id79e13f5afb1a640b5afed21a9fb50b7fc8aa5f0

Former-commit-id: e3c1af0768706f83f113a0832b0480ccd0072977