ChangesCollection: Lookup change in index when legacy ID is given
So far when a legacy change ID was specified we looked up the change
in the database (via ChangeControl#controlFor(Change.Id,CurrentUser)).
We cannot load the change via ChangeNotes.Factory since we do not know
the project. Hence lookup the change in the secondary index like we
already do it when a Change-Id is given.

ConsistencyCheckerIT creates changes manually and must now take care
to put them into the index, so that the query can find them.

Change-Id: I4885855b406c0da854ce5490ffb91b0f93808c3e
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 02417753a93b5613317c92c2aa02ab7566a8766d